requirements.txt
bpy==1.2.3
numpy==1.17.3
Pillow==6.2.1
